feat: Refactor MessagesPage with modular components and placeholders
- Split MessagesPage into reusable client-side components:
  - Added AttachmentWarning.js for attachment notifications
  - Added DetailedMessageView.js for full message display
  - Added MessageItem.js for message list items
  - Added MessageReceivers.js for recipient handling
  - Added TagsList.js for tag display
- Created lib/placeholders.js with constant MESSAGE_PLACEHOLDERS
- Simplified MessagesPage by:
  - Removing inline components
  - Importing new modular components
  - Converting to default export function
- Maintained all existing functionality with improved code organization
