+++ lib/placeholders.js
@@ -1,0 +1,7 @@
+export const MESSAGE_PLACEHOLDERS = {
+  UNKNOWN_SENDER: "Unknown sender",
+  UNKNOWN_RECEIVER: "Unknown receiver",
+  NO_TOPIC: "No topic",
+  NO_CONTENT: "No content available.",
+  SEND_DATE_UNAVAILABLE: "Send date unavailable"
+};
